clean up and update to java 21

+++ src/main/java/de/holube/fakestudy/study/category/DiscreteNumberCategory.java
@@ -7,7 +7,7 @@ public class DiscreteNumberCategory extends NumberCategory {
         super.calculate(amountSubjects);
 
         for (int i = 0; i < results.length; i++) {
-            results[i] = Math.round(results[i]);
+            results[i] = (double) Math.round(results[i]);
         }
     }
